refactor provider checks and add geometry ingest support for ogc services
