+++ eventkit_cloud/utils/services/__init__.py
@@ -3,8 +3,26 @@ import importlib
 client_type_map = {
     "arcgis-feature": "eventkit_cloud.utils.services.arcgis.ArcGIS",
     "arcgis-raster": "eventkit_cloud.utils.services.arcgis.ArcGIS",
+    "wfs": "eventkit_cloud.utils.services.wfs.WFS",
+    "wcs": "eventkit_cloud.utils.services.wcs.WCS",
+    "wms": "eventkit_cloud.utils.services.wms.WMS",
+    "osm": "eventkit_cloud.utils.services.overpass.Overpass",
+    "osm-generic": "eventkit_cloud.utils.services.overpass.Overpass",
+    "wmts": "eventkit_cloud.utils.services.wmts.WMTS",
+    "arcgis-raster": "eventkit_cloud.utils.services.arcgis.ArcGIS",
+    "arcgis-feature": "eventkit_cloud.utils.services.arcgis.ArcGIS",
+    "tms": "eventkit_cloud.utils.services.tms.TMS",
+    "vector-file": "eventkit_cloud.utils.services.file.FileClient",
+    "raster-file": "eventkit_cloud.utils.services.file.FileClient",
+    "ogcapi-process": "eventkit_cloud.utils.services.OGCAPIProcess.OGCAPIProcess",
+    "ogcapi-process-elevation": "eventkit_cloud.utils.services.OGCAPIProcess.OGCAPIProcess",
+    "ogcapi-process-raster": "eventkit_cloud.utils.services.OGCAPIProcess.OGCAPIProcess",
+    "ogcapi-process-vector": "eventkit_cloud.utils.services.OGCAPIProcess.OGCAPIProcess",
     "default": "eventkit_cloud.utils.services.base.GisClient",
 }
+
+
+DEFAULT_CACHE_TIMEOUT = 60 * 30  # 30 minutes
 
 
 def get_client(provider_type_name):
