+++ eventkit_cloud/utils/services/file.py
@@ -1,0 +1,8 @@
+import requests
+
+from eventkit_cloud.utils.services.base import GisClient
+
+
+class FileClient(GisClient):
+    def check_response(self, head_only=True) -> requests.Response:
+        super().check_response(head_only=head_only)
