+++ eventkit_cloud/utils/services/interface.py
@@ -9,7 +9,7 @@ class IGisClient(abc.ABC):
     aoi = None
 
     @abc.abstractmethod
-    def find_layer(self, root):
+    def find_layers(self, root):
         pass
 
     @abc.abstractmethod
@@ -34,4 +34,12 @@ class IGisClient(abc.ABC):
 
     @abc.abstractmethod
     def download_product_geometry(self) -> Optional[Polygon]:
+        pass
+
+    @abc.abstractmethod
+    def validate_response(self, response) -> bool:
+        pass
+
+    @abc.abstractmethod
+    def check_response(self, head_only=False) -> requests.Response:
         pass
