+++ eventkit_cloud/utils/services/ows.py
@@ -1,9 +1,23 @@
+# -*- coding: utf-8 -*-
+import logging
 import re
-from urllib.parse import urljoin
+import xml.etree.ElementTree as ET
+from io import StringIO
+from typing import Optional, List
 
 import requests
+from django.contrib.gis.geos import Polygon
 
 from eventkit_cloud.utils.services.base import GisClient
+from eventkit_cloud.utils.services.check_result import CheckResult
+from eventkit_cloud.utils.services.errors import (
+    UnsupportedFormatError,
+    MissingLayerError,
+    ServiceError,
+    ProviderCheckError,
+)
+
+logger = logging.getLogger(__name__)
 
 
 class OWS(GisClient):
@@ -23,19 +37,74 @@ class OWS(GisClient):
         self.service_url = re.sub(r"(?i)(version|service|request)=.*?(&|$)", "", self.service_url)
 
         self.layer = self.layer.lower()
+        self.layer_elements = None
 
-    def find_layer(self, root):
+    def find_layers(self, root):
         raise NotImplementedError("Method is specific to provider type")
 
-    def get_bbox(self, element):
+    def get_bbox(self, elements) -> List[float]:
         raise NotImplementedError("Method is specific to provider type")
 
     def get_layer_name(self):
         raise NotImplementedError("Method is specific to provider type")
 
-    def get_response(self, url=None) -> requests.Response:
-        url = url or self.service_url
-        service_url = url.rstrip("/\\")
-        processes_endpoint = urljoin(service_url, "processes/")
+    def validate_response(self, response: requests.Response) -> bool:
+        try:
+            xml = response.content.decode()
+            xmll = xml.lower()
+            doctype = re.search(r"<!DOCTYPE[^>[]*(\[[^]]*\])?>", xml)
+            if doctype is not None:
+                doctype_pos = doctype.end()
+                xmll = xml[:doctype_pos] + xml[doctype_pos + 1 :].lower()
 
-        return self.session.get(url=processes_endpoint, timeout=self.timeout)
+            xmll = xmll.replace("![cdata[", "![CDATA[")
+
+            # Strip namespaces from tags (from http://bugs.python.org/issue18304)
+            iterator = ET.iterparse(StringIO(xmll))
+            for event, element in iterator:
+                if "}" in element.tag:
+                    element.tag = element.tag.split("}", 1)[1]
+            root = iterator.root
+            try:
+                layer_elements = self.find_layers(root)
+            except UnsupportedFormatError:
+                logger.error("Missing expected root layer", exc_info=True)
+                raise ProviderCheckError(CheckResult.UNKNOWN_FORMAT)
+            except MissingLayerError:
+                logger.error("Missing expected layer %s", self.layer, exc_info=True)
+                raise ProviderCheckError(CheckResult.LAYER_NOT_AVAILABLE)
+            except ServiceError:
+                logger.error("Failed to properly parse the response", exc_info=True)
+                logger.info(xml)
+                raise ProviderCheckError(CheckResult.UNKNOWN_ERROR)
+
+            if not layer_elements:
+                return False
+            self.layer_elements = layer_elements
+            geom = self.download_geometry()
+            if geom is not None:
+                self.check_intersection(geom)
+            return True
+
+        except ET.ParseError as ex:
+            logger.error("Provider check failed to parse GetCapabilities XML: {}".format(str(ex)))
+            raise ProviderCheckError(CheckResult.UNKNOWN_FORMAT)
+
+    def get_bboxes(self, elements: List[ET.Element]) -> List[List[float]]:
+        bboxes = []
+        for element in elements:
+            bbox = self.get_bbox(element)
+            if bbox:
+                bboxes.append(bbox)
+        return bboxes
+
+    def download_geometry(self) -> Optional[Polygon]:
+        if not self.layer_elements:
+            self.check(self.aoi)
+        bboxes = self.get_bboxes(self.layer_elements)
+        polygon = Polygon()
+        for bbox in bboxes:
+            polygon = polygon.union(Polygon.from_bbox(bbox))
+        if polygon.area:
+            return polygon
+        return None
