+++ eventkit_cloud/utils/services/wcs.py
@@ -1,15 +1,15 @@
 from __future__ import annotations
-import re
+
 import logging
-from typing import Union, List
+import re
+from typing import List
+from typing import TYPE_CHECKING
 
 from eventkit_cloud.utils.services.errors import MissingLayerError, UnsupportedFormatError
 from eventkit_cloud.utils.services.ows import OWS
-from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from eventkit_cloud.utils.services.provider_check import CheckResult
-
+    pass
 
 logger = logging.getLogger(__name__)
 
@@ -19,7 +19,7 @@ class WCS(OWS):
         super(WCS, self).__init__(*args, **kwargs)
         self.query["SERVICE"] = "WCS"
 
-    def find_layer(self, root) -> Union[CheckResult, List[str]]:
+    def find_layers(self, root) -> List[str]:
         """
         :param root: Name of layer to find
         :return: XML 'Layer' Element, or None if not found
@@ -36,6 +36,7 @@ class WCS(OWS):
         if not cover_names:  # No coverages are offered
             raise MissingLayerError(f"WCS Layer {self.layer} not found because no coverages were found")
 
+        coverages = None
         try:
             coverages = self.config.get("service", dict()).get("coverages")
             coverages = coverages.split(",") if coverages else None
@@ -53,27 +54,23 @@ class WCS(OWS):
 
         return covers
 
-    def get_bbox(self, elements):
-        bboxes = []
-        for element in elements:
-            envelope = element.find("lonlatenvelope")
-            if envelope is None:
-                continue
+    def get_bbox(self, element):
+        envelope = element.find("lonlatenvelope")
+        if envelope is None:
+            return
 
-            pos = list(envelope)
-            # Make sure there aren't any surprises
-            coord_pattern = re.compile(r"^-?\d+(\.\d+)? -?\d+(\.\d+)?$")
-            if not pos or not all("pos" in p.tag and re.match(coord_pattern, p.text) for p in pos):
-                continue
+        pos = list(envelope)
+        # Make sure there aren't any surprises
+        coord_pattern = re.compile(r"^-?\d+(\.\d+)? -?\d+(\.\d+)?$")
+        if not pos or not all("pos" in p.tag and re.match(coord_pattern, p.text) for p in pos):
+            return
 
-            x1, y1 = list(map(float, pos[0].text.split(" ")))
-            x2, y2 = list(map(float, pos[1].text.split(" ")))
+        x1, y1 = list(map(float, pos[0].text.split(" ")))
+        x2, y2 = list(map(float, pos[1].text.split(" ")))
 
-            minx, maxx = sorted([x1, x2])
-            miny, maxy = sorted([y1, y2])
-
-            bboxes.append([minx, miny, maxx, maxy])
-        return bboxes if bboxes else None
+        minx, maxx = sorted([x1, x2])
+        miny, maxy = sorted([y1, y2])
+        return [minx, miny, maxx, maxy]
 
     def get_layer_name(self):
         raise NotImplementedError("Method is specific to provider type")
